fix(tdd-rewrite/linux-L5): remove dead seccomp param and widen pid fields to u64
- seccomp.rs: drop unused `_prev_offset: u64` parameter from
  `count_filter_chain` and update the single call site accordingly;
  widen `SeccompInfo.pid` from u32 to u64, removing the `as u32` cast
- capabilities.rs: widen `ProcessCapabilities.pid` from u32 to u64,
  removing the `as u32` cast in `read_process_caps`
- cgroups.rs: widen `CgroupInfo.pid` from u32 to u64 and update
  `is_suspicious_cgroup` signature to accept u64; remove both `as u32`
  casts in `walk_cgroups`

+++ crates/memf-linux/src/capabilities.rs
@@ -42,7 +42,7 @@ const SUSPICIOUS_CAPS: &[(u64, &str)] = &[
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct ProcessCapabilities {
     /// Process ID.
-    pub pid: u32,
+    pub pid: u64,
     /// Process command name.
     pub name: String,
     /// Bitmask of effective capabilities.
@@ -158,7 +158,7 @@ fn read_process_caps<P: PhysicalMemoryProvider>(
     let (is_suspicious, suspicious_caps) = classify_capabilities(effective, uid);
 
     Ok(ProcessCapabilities {
-        pid: proc.pid as u32,
+        pid: proc.pid,
         name: proc.comm.clone(),
         effective,
         permitted,
